correct typo while showing candidates stats

+++ preprocessing/reduce_wikidata_size.py
@@ -21,26 +21,27 @@ def reduce_size(wikidata, ents_vocab, ents_oov):
                 
 def main():
     
-    ents_vocab = pkl.load(open("../vocabs/entities_vocab.pkl", "rb"))
-    ents_oov = pkl.load(open("../vocabs/entities_oov.pkl", "rb"))
+    if not os.path.exists("vocabs"):
+        exit("Create the vocabularies first")
+    
+    ents_vocab = pkl.load(open("vocabs/entities_vocab.pkl", "rb"))
+    ents_oov = pkl.load(open("vocabs/entities_oov.pkl", "rb"))
     
     #process wikidata1
-    wikidata = load_wikidata("../datasets/wikidata_dir/wikidata_short_1.json")
+    wikidata = load_wikidata("datasets/wikidata_dir/wikidata_short_1.json")
     wikidata = reduce_size(wikidata, ents_vocab, ents_oov)
     
-    with open('../datasets/wikidata_dir/wikidata_short_1_reduced.json', 'w', encoding='utf-8') as f:
+    with open('datasets/wikidata_dir/wikidata_short_1_reduced.json', 'w', encoding='utf-8') as f:
         json.dump(wikidata, f)
         
     #process wikidata2
-    wikidata = load_wikidata("../datasets/wikidata_dir/wikidata_short_2.json")
+    wikidata = load_wikidata("datasets/wikidata_dir/wikidata_short_2.json")
     wikidata = reduce_size(wikidata, ents_vocab, ents_oov)
     
-    with open('../datasets/wikidata_dir/wikidata_short_2_reduced.json', 'w', encoding='utf-8') as f:
+    with open('datasets/wikidata_dir/wikidata_short_2_reduced.json', 'w', encoding='utf-8') as f:
         json.dump(wikidata, f)
     
 
 if __name__ == "__main__":
     main()
 
-    
-
